# server/docs/openapi.yaml
openapi: 3.0.0
info:
  title: Ticket Manager API
  description: REST API for ticket management using ITSM data schema
  version: 2.0.0
  contact:
    name: API Support
    email: support@example.com
servers:
  - url: http://localhost:3000
    description: Development server
paths:
...
  /api/tickets:
    get:
      summary: List tickets
      description: Retrieves a list of tickets with optional filtering, pagination, and sorting
      tags:
        - Tickets
      security:
        - BearerAuth: []
      parameters:
        - in: query
          name: status
          schema:
            type: string
          description: Filter by ticket status (e.g., Open, Closed)
        - in: query
          name: priority
          schema:
            type: string
          description: Filter by priority level (1, 2, 3, 4, 5, NS, NA)
        - in: query
          name: category
          schema:
            type: string
          description: Filter by category (e.g., incident, request for information)
        - in: query
          name: id
          schema:
            type: integer
          description: Filter by ticket ID (e.g., 4)
        - in: query
          name: search
          schema:
            type: string
          description: Search in title and description
        - in: query
          name: limit
          schema:
            type: integer
            default: 100
            maximum: 1000
          description: Maximum number of results
        - in: query
          name: offset
          schema:
            type: integer
            default: 0
          description: Number of records to skip (pagination)
        - in: query
          name: sortBy
          schema:
            type: string
            default: created_at
            enum: [id, title, status, priority, category, created_at, updated_at]
          description: Field to sort by
        - in: query
          name: sortOrder
          schema:
            type: string
            default: DESC
            enum: [ASC, DESC]
          description: Sort order
      responses:
        '200':
          description: List of tickets with pagination info
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    type: array
                    items:
                      $ref: '#/components/schemas/Ticket'
                  pagination:
                    type: object
                    properties:
                      total:
                        type: integer
                        example: 150
                      limit:
                        type: integer
                        example: 10
                      offset:
                        type: integer
                        example: 0
                      hasMore:
                        type: boolean
                        example: true
    post:
      summary: Create ticket
      description: Creates a new ticket
      tags:
        - Tickets
      security:
        - BearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - title
              properties:
                title:
                  type: string
                  minLength: 1
                  maxLength: 255
                  example: SUB000501
                  description: Ticket title (from CI_Name in ITSM data)
                description:
                  type: string
                  maxLength: 5000
                  example: subapplication
                  description: Ticket description (from CI_Cat in ITSM data)
                priority:
                  type: string
                  default: '3'
                  example: '4'
                  description: Priority level (1, 2, 3, 4, 5, NS, NA)
                category:
                  type: string
                  default: incident
                  example: incident
                  description: Ticket category
                impact:
                  type: string
                  example: '3'
                  description: Impact level (1, 2, 3, 4, 5, NS)
                urgency:
                  type: string
                  example: '3'
                  description: Urgency level (1, 2, 3, 4, 5, NS)
      responses:
        '201':
          description: Ticket created successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Ticket'
        '400':
          description: Validation error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
...
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: Token
  schemas:
...
    Ticket:
      type: object
      properties:
        id:
          type: integer
          example: 4
          description: Unique ticket identifier from ITSM data
        title:
          type: string
          example: SUB000501
          description: Ticket title (from CI_Name in ITSM data)
        description:
          type: string
          example: subapplication
          description: Ticket description (from CI_Cat in ITSM data)
        status:
          type: string
          example: Open
          description: Ticket status (Open, Closed)
        priority:
          type: string
          example: '4'
          description: Priority level (1, 2, 3, 4, 5, NS, NA)
        category:
          type: string
          example: incident
          description: Ticket category
        impact:
          type: string
          nullable: true
          example: '4'
          description: Impact level (1, 2, 3, 4, 5, NS)
        urgency:
          type: string
          nullable: true
          example: '3'
          description: Urgency level (1, 2, 3, 4, 5, NS)
        created_at:
          type: string
          format: date-time
          nullable: true
          example: 2013-08-21T11:20:00.000Z
          description: When the ticket was opened (from Open_Time)
        resolved_at:
          type: string
          format: date-time
          nullable: true
          example: 2013-10-01T16:38:00.000Z
          description: When the ticket was resolved (from Resolved_Time)
        closed_at:
          type: string
          format: date-time
          nullable: true
          example: 2013-10-01T16:38:00.000Z
          description: When the ticket was closed (only present when status is Closed)
        updated_at:
          type: string
          format: date-time
          example: 2013-10-01T16:38:00.000Z
      description: Ticket object with conditional field visibility
    Error:
      type: object
      properties:
        error:
          type: string
          example: Validation Error
        message:
          type: string
          example: Title is required
...